Remove data and news specific `errors` modules
The three different `errors` modules were a bit redundant and confusing. This moves them all to a central place (which involves more deleting duplicated stuff than actually moving things).
