+++ covid19_sfbayarea/errors.py
@@ -1,3 +1,6 @@
+import requests
+
+
 class FormatError(Exception):
     """
     A custom error to raise whenever a scraper runs into something in an
@@ -12,3 +15,12 @@ class PowerBiQueryError(ValueError):
     Represents an error returned by PowerBI in response to a query.
     """
     ...
+
+
+class BadRequest(requests.exceptions.HTTPError):
+    """
+    Represents a detailed error message from a web server.
+    """
+    def __str__(self) -> str:
+        message = super().__str__()
+        return f'{message} (status: {self.response.status_code})'
