--- covid19_sfbayarea/news/errors.py
@@ -1,7 +1,0 @@
-class FormatError(Exception):
-    """
-    A custom error to raise whenever a scraper runs into something in an
-    unexpected format. This usually means that the website the scraper is
-    accessing has changed.
-    """
-    pass
