+++ covid19_sfbayarea/news/sonoma.py
@@ -1,9 +1,9 @@
 from bs4 import BeautifulSoup, element  # type: ignore
 from typing import List
 from urllib.parse import urljoin
+from ..errors import FormatError
 from ..utils import parse_datetime
 from .base import NewsScraper
-from .errors import FormatError
 from .feed import NewsItem
 from .utils import get_base_url, is_covid_related, normalize_whitespace
